+++ clang-tools-extra/clang-tidy/bsl/BslTidyModule.cpp
@@ -70,6 +70,7 @@
 #include "UserDefinedTypeNamesMatchHeaderNameCheck.h"
 #include "UsingDeclScopeCheck.h"
 #include "UsingIdentUniqueNamespaceCheck.h"
+#include "UsingNamespaceForbiddenCheck.h"
 #include "VarBracedInitCheck.h"
 
 namespace clang {
@@ -201,6 +202,8 @@ public:
         "bsl-using-decl-scope");
     CheckFactories.registerCheck<UsingIdentUniqueNamespaceCheck>(
         "bsl-using-ident-unique-namespace");
+    CheckFactories.registerCheck<UsingNamespaceForbiddenCheck>(
+        "bsl-using-namespace-forbidden");
     CheckFactories.registerCheck<VarBracedInitCheck>(
         "bsl-var-braced-init");
   }