+++ clang-tools-extra/test/clang-tidy/checkers/bsl-using-namespace-forbidden.cpp
@@ -1,0 +1,10 @@
+// RUN: %check_clang_tidy %s bsl-using-namespace-forbidden %t
+
+namespace bsl
+{}
+
+namespace
+{}
+
+using namespace bsl;
+// CHECK-MESSAGES: :[[@LINE-1]]:1: warning: using directives are forbidden [bsl-using-namespace-forbidden]